+++ src/config/main.rs
@@ -2,7 +2,11 @@ use std::path::PathBuf;
 
 use log::{Level as LogLevel};
 use rsa::RsaPrivateKey;
-use serde::Deserialize;
+use serde::{
+    Deserialize,
+    Deserializer,
+    de::Error as DeserializerError,
+};
 use url::Url;
 
 use crate::activitypub::constants::ACTOR_KEY_SUFFIX;
@@ -13,6 +17,36 @@ use crate::utils::urls::guess_protocol;
 use super::blockchain::BlockchainConfig;
 use super::environment::Environment;
 use super::MITRA_VERSION;
+
+#[derive(Clone, PartialEq)]
+pub enum RegistrationType {
+    Open,
+    Invite,
+}
+
+impl Default for RegistrationType {
+    fn default() -> Self { Self::Invite }
+}
+
+impl<'de> Deserialize<'de> for RegistrationType {
+    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+        where D: Deserializer<'de>
+    {
+        let registration_type_str = String::deserialize(deserializer)?;
+        let registration_type = match registration_type_str.as_str() {
+            "open" => Self::Open,
+            "invite" => Self::Invite,
+            _ => return Err(DeserializerError::custom("unknown registration type")),
+        };
+        Ok(registration_type)
+    }
+}
+
+#[derive(Clone, Default, Deserialize)]
+pub struct RegistrationConfig {
+    #[serde(rename = "type")]
+    pub registration_type: RegistrationType,
+}
 
 fn default_log_level() -> LogLevel { LogLevel::Info }
 
@@ -52,8 +86,10 @@ pub struct Config {
     #[serde(skip)]
     pub(super) instance_rsa_key: Option<RsaPrivateKey>,
 
+    pub(super) registrations_open: Option<bool>, // deprecated
+
     #[serde(default)]
-    pub registrations_open: bool, // default is false
+    pub registration: RegistrationConfig,
 
     // EIP-4361 login message
     #[serde(default = "default_login_message")]