+++ src/main.rs
@@ -39,9 +39,12 @@ use mitra::web_client::views as web_client;
 
 #[actix_web::main]
 async fn main() -> std::io::Result<()> {
-    let config = parse_config();
+    let (config, config_warnings) = parse_config();
     configure_logger(config.log_level);
     log::info!("config loaded from {}", config.config_path);
+    for warning in config_warnings {
+        log::warn!("{}", warning);
+    };
 
     let db_pool = create_pool(&config.database_url);
     let mut db_client = get_database_client(&db_pool).await.unwrap();